+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Security and Quality - Uncontrolled data used in path expression - routes/scan.js #9
 - Security and Quality - Uncontrolled data used in path expression - utils/fileParser.js #10 / #11 / #12
 - Correction - Load vue History not refresh,
+- Correction - Report UI5 - Display End of Maintenance.
 
 ## v1.2.0 (2026-06-20)
 - Add documentation gh-pages,